requirements.txt
# Core dependencies
pdf2image>=1.16.3
pillow>=9.0.0
requests>=2.28.0
openai>=1.0.0

# PDF processing
PyPDF2>=3.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# OCR & Vision
pdf2image>=1.16.3

# Web server
Flask>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0

# Utilities
python-dotenv>=0.19.0

# Cloud OCR APIs
# 豆包 (Doubao) - 推荐用于医学检验报告
# 官网: https://www.volcengine.com/product/doubao
# 需要注册并获取API Key

# Kimi
# 官网: https://platform.moonshot.cn/
# 需要注册并获取API Key

# 阿里云OCR
# 官网: https://www.aliyun.com/product/ai/ocr
# 需要注册并获取AppCode
